requirements.txt
Pillow==10.0.1
pynput==1.7.6
